Pair-programming session to combine queries,  and populating database scripts.
Removed old scripts.
