+++ .gitignore
@@ -13,6 +13,8 @@ _migrations/__pycache__/*
 /Ploutos/Ploutos/dashboard/pycashe/*
 *.sqlite3
 
+
+
 #ignore migrations except the folder and __init__.py
 **/migrations/**
 !**/migrations
